requirements.txt
Flask==3.0.0
pandas==2.2.1
scikit-learn
python-docx==1.1.0
spacy==3.7.4